Add ANY type support.
Parameters in a DECLARE statement can be declared of type ANY. This is
because in QB, declare statements should precede all other statements,
including user defined type definitions. That means, when the declare
statement is encountered, user-defined types are not available and the
only way to have them is to declare them as type ANY so that type
checking for that parameter will be disabled.